requirements.txt
ultralytics>=8.3.0
opencv-python>=4.10.0
numpy>=1.26.0
pygame>=2.6.0
llama-cpp-python>=0.3.2
huggingface-hub>=0.23.0
